separate widgets of different concerns
