+++ lib/domain/models/onboarding_view_models.dart
@@ -1,0 +1,11 @@
+class SliderObject {
+  final String title;
+  final String subTitle;
+  final String imageUrl;
+
+  SliderObject({
+    required this.title,
+    required this.subTitle,
+    required this.imageUrl,
+  });
+}
