+++ lib/presentation/onboarding/view_model/onboarding_view_model.dart
@@ -2,11 +2,11 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_store_app/presentation/base/base_view_model.dart';
-import 'package:flutter_store_app/presentation/onboarding/veiw/onboarding_view.dart';
 import 'package:flutter_store_app/presentation/onboarding/view_model/onboarding_view_model_state.dart';
 import 'package:flutter_store_app/resources/constants_dart.dart';
 import 'package:flutter_store_app/resources/routes_manager.dart';
 
+import '../../../domain/models/onboarding_view_models.dart';
 import '../../../resources/assets_manager.dart';
 import '../../../resources/string_manager.dart';
